Initial commit.
This commit contains the initial version of Findis. It consists of a
solution containing 4 projects:

Findis.Business: Contains the main business rules, includes the database
logic and has a public interface that can be used for interaction with
the system.

Findis.Common: Contains classes that are shared across the solution.

Findis.Proto: A prototype user interface that was quickly put together
in order to test the system.

Findis.Test: A project for unit tests for the entire solution. Currently
only the publich managers in the Business layer have been tested.